Split up models, panels and snippets in great domestic
