+++ export_readiness/views.py
@@ -1,11 +1,11 @@
 from rest_framework.generics import RetrieveAPIView
 
 from conf.signature import SignatureCheckPermission
-from export_readiness import models, serializers
+from export_readiness import serializers, snippets
 
 
 class PageLookupByTagListAPIEndpoint(RetrieveAPIView):
-    queryset = models.Tag
+    queryset = snippets.Tag
     serializer_class = serializers.TagSerializer
     lookup_field = 'slug'
     permission_classes = [SignatureCheckPermission]
